requirements.txt
chromadb
langchain
sentence_transformers
lark
tqdm
python-dotenv
openai
streamlit